+++ web/app/(commonLayout)/manager/basic-content.tsx
